requirements.txt
langchain==0.3.0
langchain-text-splitters==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-community==0.3.0
chromadb==0.5.0
python-dotenv==1.0.1
loguru==0.7.2
sentence-transformers==3.3.1
rank-bm25==0.2.2
torch==2.5.1